chore: move types to its own folder
